utils/openai_client.py
import os
from langchain_openai import AzureChatOpenAI
from dotenv import load_dotenv

load_dotenv()

# Centralized LLM instance for the entire application
llm = AzureChatOpenAI(
    azure_endpoint=os.getenv("AZURE_OPENAI_ENDPOINT"),
    api_key=os.getenv("AZURE_OPENAI_API_KEY"),
    api_version=os.getenv("AZURE_OPENAI_API_VERSION", "2024-02-15-preview"),
    deployment_name=os.getenv("AZURE_OPENAI_DEPLOYMENT", "gpt-4o-mini"),
    temperature=0,
    max_tokens=1500,
)

