requirements.txt
pandas
streamlit
openai
matplotlib
seaborn
numpy
scikit-learn
langchain
langchain-openai
langchain-core
python-dotenv
plotly
lxml
openpyxl
scipy
